+++ src/ffmpeg/libavcodec/golomb.h
@@ -31,9 +31,8 @@
 #define AVCODEC_GOLOMB_H
 
 #include <stdint.h>
-
-#include "get_bits.h"
 #include "libavutil/internal.h"
+#include "get_bits.h"
 
 #define INVALID_VLC           0x80000000
 
@@ -480,7 +479,7 @@ static inline int get_ur_golomb_jpegls(GetBitContext *gb, int k, int limit,
     } else {
         int i;
         for (i = 0; i + MIN_CACHE_BITS <= limit && SHOW_UBITS(re, gb, MIN_CACHE_BITS) == 0; i += MIN_CACHE_BITS) {
-            if (gb->size_in_bits <= (int)re_index) {
+            if (gb->size_in_bits <= re_index) {
                 CLOSE_READER(re, gb);
                 return -1;
             }
